update imports to use information matrix utils. Change uppercase names to lowercase
