+++ sysidentpy/tests/test_narmax_base.py
@@ -2,23 +2,11 @@ from typing import Optional
 
 import numpy as np
 import pytest
-from unittest.mock import MagicMock
 from numpy.testing import (
     assert_almost_equal,
     assert_array_equal,
     assert_equal,
     assert_raises,
-)
-
-from sysidentpy.utils.simulation import (
-    get_index_from_regressor_code,
-    list_output_regressor_code,
-    list_input_regressor_code,
-)
-
-from sysidentpy.utils.lags import (
-    get_lag_from_regressor_code,
-    get_max_lag_from_model_code,
 )
 
 from sysidentpy.basis_function import Polynomial, Fourier
@@ -31,14 +19,7 @@ from sysidentpy.parameter_estimation.estimators import (
 from sysidentpy.narmax_base import (
     house,
     rowhouse,
-    shift_column,
-    _create_lagged_X,
-    _create_lagged_y,
 )
-
-from sysidentpy.utils.lags import _process_ylag, _process_xlag
-
-from sysidentpy.utils.generate_data import get_miso_data, get_siso_data
 
 GR = RegressorDictionary()
 bf_polynomial = Polynomial(degree=2)
@@ -256,72 +237,6 @@ def test_row_house():
     assert_almost_equal(rowhouse(a, b), output)
 
 
-def test_get_index_from_regressor_code():
-    model = np.array(
-        [
-            [1001, 0],  # y(k-1)
-            [2001, 1001],  # x1(k-1)y(k-1)
-            [2002, 0],  # x1(k-2)
-        ]
-    )
-
-    regressor_space = np.array(
-        [
-            [0, 0],
-            [1001, 0],
-            [2001, 0],
-            [2002, 0],
-            [1001, 1001],
-            [2001, 1001],
-            [2002, 1001],
-            [2001, 2001],
-            [2002, 2001],
-            [2002, 2002],
-        ]
-    )
-    index = get_index_from_regressor_code(
-        regressor_code=regressor_space, model_code=model
-    )
-
-    assert (index == np.array([1, 3, 5])).all()
-
-
-def test_list_output_regressor():
-    model = np.array(
-        [
-            [1001, 0],  # y(k-1)
-            [2001, 1001],  # x1(k-1)y(k-1)
-            [2002, 0],  # x1(k-2)
-        ]
-    )
-
-    y_code = list_output_regressor_code(model)
-    assert (y_code == np.array([1001, 1001])).all()
-
-
-def test_list_input_regressor():
-    model = np.array(
-        [
-            [1001, 0],  # y(k-1)
-            [2001, 1001],  # x1(k-1)y(k-1)
-            [2002, 0],  # x1(k-2)
-        ]
-    )
-
-    x_code = list_input_regressor_code(model)
-    assert (x_code == np.array([2001, 2002])).all()
-
-
-def test_get_lag_from_regressor_code():
-    list_regressor1 = np.array([2001, 2002])
-    list_regressor2 = np.array([1004, 1002])
-    max_lag1 = get_lag_from_regressor_code(list_regressor1)
-    max_lag2 = get_lag_from_regressor_code(list_regressor2)
-
-    assert max_lag1 == 2
-    assert max_lag2 == 4
-
-
 def test_get_max_lag():
     output1 = 1
     r = RegressorDictionary(
@@ -333,69 +248,6 @@ def test_get_max_lag():
     )._get_max_lag()
     assert_equal(output1, r)
     assert_equal(output2, r2)
-
-
-def test_shift_column():
-    a = np.array(
-        [
-            0.42544384,
-            0.39365905,
-            0.22209413,
-            0.69760074,
-            0.88183369,
-            0.24818225,
-            0.78482346,
-            0.26967285,
-            0.53987842,
-            0.17367185,
-        ]
-    ).reshape(-1, 1)
-
-    output = np.array(
-        [
-            [0.0],
-            [0.0],
-            [0.42544384],
-            [0.39365905],
-            [0.22209413],
-            [0.69760074],
-            [0.88183369],
-            [0.24818225],
-            [0.78482346],
-            [0.26967285],
-        ]
-    )
-    r = shift_column(a, 2)
-    assert_almost_equal(output, r)
-
-
-def test_process_xlag():
-    a = np.array(
-        [
-            0.42544384,
-            0.39365905,
-            0.22209413,
-            0.69760074,
-            0.88183369,
-            0.24818225,
-            0.78482346,
-            0.26967285,
-            0.53987842,
-            0.17367185,
-        ]
-    ).reshape(-1, 1)
-
-    n_inputs, xlag = _process_xlag(a.reshape(-1, 1), xlag=2)
-    output1 = 1
-    output2 = list(range(1, 3))
-    assert_equal(output1, n_inputs)
-    assert_equal(output2, xlag)
-
-
-def test_process_ylag():
-    ylag = _process_ylag(ylag=2)
-    output1 = list(range(1, 3))
-    assert_equal(output1, ylag)
 
 
 def test_errors():
@@ -467,76 +319,6 @@ def test_regressor_space_raise():
     )
 
 
-def test_model_information_get_lag():
-    laglist = np.array([2001, 2002, 3001, 3002, 1001, 1002])
-    output = 2
-    r1 = get_lag_from_regressor_code(laglist)
-    assert r1 == output
-
-
-def test_model_information_empty_list():
-    laglist = np.array([])
-    output = 1
-    r1 = get_lag_from_regressor_code(laglist)
-    assert r1 == output
-
-
-def test_get_max_lag_from_model_code():
-    model = np.array(
-        [
-            [1001, 0],  # y(k-1)
-            [2001, 1001],  # x1(k-1)y(k-1)
-            [2002, 0],  # x1(k-2)
-        ]
-    )
-    assert get_max_lag_from_model_code(model) == 2
-
-
-def test_process_lag():
-    x_train, _, _, _ = get_miso_data(
-        n=10, colored_noise=False, sigma=0.001, train_percentage=90
-    )
-    assert_raises(ValueError, _process_xlag, X=x_train, xlag=2)
-
-
-def test_process_lag_n1():
-    x_train, _, _, _ = get_siso_data(
-        n=10, colored_noise=False, sigma=0.001, train_percentage=90
-    )
-    n_inputs, xlag = _process_xlag(X=x_train, xlag=2)
-    assert n_inputs == 1
-    assert list(xlag) == [1, 2]
-
-
-def test_create_lagged_x():
-    X = np.array([1, 2, 3, 4, 5, 6]).reshape(-1, 1)
-    r = _create_lagged_X(X=X, n_inputs=1, xlag=[1, 2])
-    assert_equal(
-        r,
-        np.array(
-            [[0.0, 0.0], [1.0, 0.0], [2.0, 1.0], [3.0, 2.0], [4.0, 3.0], [5.0, 4.0]]
-        ),
-    )
-
-
-def test_create_lagged_x_miso():
-    X = np.array(range(1, 13)).reshape(-1, 2)
-    r = _create_lagged_X(X=X, n_inputs=2, xlag=[[1, 2], [1, 2]])
-    assert_equal(
-        r,
-        np.array(
-            [
-                [0.0, 0.0, 0.0, 0.0],
-                [1.0, 0.0, 2.0, 0.0],
-                [3.0, 1.0, 4.0, 2.0],
-                [5.0, 3.0, 6.0, 4.0],
-                [7.0, 5.0, 8.0, 6.0],
-                [9.0, 7.0, 10.0, 8.0],
-            ]
-        ),
-    )
-
-
 def test_model_predict():
     model = FROLS(
         n_terms=5,
@@ -604,7 +386,7 @@ def test_model_predict_fourier_steps_none():
         basis_function=Fourier(degree=2, n=1),
     )
     model.fit(X=X_train, y=y_train)
-    yhat = model._basis_function_predict(X=X_test, y_initial=y_test)
+    yhat = model._basis_function_predict(x=X_test, y_initial=y_test)
     assert_almost_equal(yhat.mean(), y_test[model.max_lag : :].mean(), decimal=6)
 
 
@@ -668,7 +450,7 @@ def test_model_predict_fourier_value_error():
     assert_raises(
         ValueError,
         model._basis_function_n_step_prediction,
-        X=X_test,
+        x=X_test,
         y=y_test,
         steps_ahead=1,
         forecast_horizon=None,
@@ -690,7 +472,7 @@ def test_model_predict_fourier_horizon_error():
     assert_raises(
         ValueError,
         model._basis_function_n_steps_horizon,
-        X=X_test,
+        x=X_test,
         y=y_test,
         steps_ahead=1,
         forecast_horizon=10,
